refactor: bugfix - identifiers type

+++ src/main/java/com/example/carsharing/service/CarService.java
@@ -4,11 +4,11 @@ import com.example.carsharing.model.Car;
 import java.util.List;
 
 public interface CarService {
-    Car getById(Integer id);
+    Car getById(Long id);
 
     Car save(Car car);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     List<Car> getAll();
 }
